# src/main/resources/static/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /v1/product/{id}:
    get:
      tags:
      - Products Inquiry
      description: Get Product
      operationId: getProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden Access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: Internal server error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      tags:
      - Update product
      description: Update product
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden Access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: Internal server error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      tags:
      - delete product
      description: Delete product
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden Access
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "404":
          description: Resource Not Found
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "500":
          description: Internal server error occurred
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  /v1/product:
    get:
      tags:
      - All Products Inquiry
      description: Get All Products
      operationId: getAllProducts
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden Access
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "404":
          description: Resource Not Found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "500":
          description: Internal server error occurred
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags:
      - Add Product
      description: Create new product
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Forbidden Access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: Internal server error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
components:
  schemas:
    ProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        productType:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 1
        unitPrice:
          type: number
          format: double
          minimum: 0
      required:
      - description
      - name
      - productType
      - quantity
      - unitPrice
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        productType:
          type: string
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
          format: double